Refactor TTLMonitor module to cbm
