+++ src/CMakeLists.txt
@@ -71,7 +71,7 @@ get_filename_component(DREAM_CFG_FILE "${ESS_MODULE_DIR}/dream/configs/DreamEndc
 get_filename_component(MAGIC_CFG_FILE "${ESS_MODULE_DIR}/dream/configs/MagicInst.json" ABSOLUTE)
 get_filename_component(FREIA_CFG_FILE "${ESS_MODULE_DIR}/freia/configs/freia.json" ABSOLUTE)
 get_filename_component(TREX_CFG_FILE "${ESS_MODULE_DIR}/trex/configs/trex.json" ABSOLUTE)
-get_filename_component(TTLMON_CFG_FILE "${ESS_MODULE_DIR}/ttlmonitor/configs/freiamon.json" ABSOLUTE)
+get_filename_component(CBM_CFG_FILE "${ESS_MODULE_DIR}/cbm/configs/freiamon.json" ABSOLUTE)
 get_filename_component(NMX_CFG_FILE "${ESS_MODULE_DIR}/nmx/configs/nmx.json" ABSOLUTE)
 get_filename_component(LOKI_CFG_FILE "${ESS_MODULE_DIR}/loki/configs/LokiFull.json" ABSOLUTE)
 get_filename_component(LOKI_CAL_FILE "${ESS_MODULE_DIR}/loki/configs/lokinullcalib.json" ABSOLUTE)
@@ -90,7 +90,7 @@ add_custom_target(copycfgs
     COMMAND ${CMAKE_COMMAND} -E copy ${MAGIC_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
     COMMAND ${CMAKE_COMMAND} -E copy ${FREIA_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
     COMMAND ${CMAKE_COMMAND} -E copy ${TREX_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
-    COMMAND ${CMAKE_COMMAND} -E copy ${TTLMON_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
+    COMMAND ${CMAKE_COMMAND} -E copy ${CBM_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
     COMMAND ${CMAKE_COMMAND} -E copy ${NMX_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
     COMMAND ${CMAKE_COMMAND} -E copy ${LOKI_CFG_FILE} ${PROJECT_BINARY_DIR}/configs/
     COMMAND ${CMAKE_COMMAND} -E copy ${LOKI_CAL_FILE} ${PROJECT_BINARY_DIR}/configs/
@@ -109,13 +109,13 @@ add_custom_target(runefu
     COMMAND bifrost  "-f" "../configs/bifrost.json"    "--calibration" "../configs/bifrostnullcalib.json"  "-s" "2" "--nohwcheck" "--dumptofile" "deleteme_user.h5"
     COMMAND miracles "-f" "../configs/miracles.json"   "--calibration" "../configs/miraclesnullcalib.json" "-s" "2" "--nohwcheck"
     COMMAND cspec    "-f" "../configs/cspec.json"      "--calibration" "../configs/cspecnullcalib.json"    "-s" "2" "--nohwcheck" "--dumptofile" "deleteme_cspec.h5"
-    COMMAND perfgen "-s" "1" "--nohwcheck"
+    COMMAND perfgen  "-s" "1" "--nohwcheck"
     COMMAND dream    "-f" "../configs/DreamEndcap.json" "-s" "2" "--nohwcheck"
     COMMAND magic    "-f" "../configs/MagicInst.json" "-s" "2" "--nohwcheck"
     COMMAND freia    "-f" "../configs/freia.json" "-s" "2" "--nohwcheck" "-l" "Info" "--log_file" "deleteme_graylog.log"
     COMMAND trex     "-f" "../configs/trex.json" "-s" "2" "--nohwcheck"
-    COMMAND ttlmonitor "-f" "../configs/freiamon.json" "-s" "2" "--nohwcheck" "--dumptofile" "deleteme_ttlmon_" "-t" "freia_beam_monitor"
-    COMMAND nmx "-f" "../configs/nmx.json" "-s" "2"  "--nohwcheck"
+    COMMAND cbm      "-f" "../configs/freiamon.json" "-s" "2" "--nohwcheck" "--dumptofile" "deleteme_ttlmon_" "-t" "freia_beam_monitor"
+    COMMAND nmx      "-f" "../configs/nmx.json" "-s" "2"  "--nohwcheck"
     COMMAND timepix3 "-f" "../configs/timepix3.json" "-s" "2" "--nohwcheck"
-    DEPENDS copycfgs dream loki bifrost miracles freia trex ttlmonitor nmx cspec timepix3
+    DEPENDS copycfgs dream loki bifrost miracles freia trex cbm nmx cspec timepix3
     )
