+++ src/modules/CMakeLists.txt
@@ -16,7 +16,7 @@ add_subdirectory(loki)
 add_subdirectory(miracles)
 add_subdirectory(nmx)
 add_subdirectory(perfgen)
-add_subdirectory(ttlmonitor)
+add_subdirectory(cbm)
 
 
 #=============================================================================
