+++ src/modules/cbm/TTLMonitorBase.h
@@ -12,7 +12,7 @@
 #include <common/detector/Detector.h>
 #include <common/kafka/EV44Serializer.h>
 #include <memory>
-#include <ttlmonitor/Counters.h>
+#include <cbm/Counters.h>
 
 namespace TTLMonitor {
