+++ src/modules/cbm/geometry/Parser.cpp
@@ -9,7 +9,7 @@
 //===----------------------------------------------------------------------===//
 
 #include <common/debug/Trace.h>
-#include <ttlmonitor/geometry/Parser.h>
+#include <cbm/geometry/Parser.h>
 
 namespace TTLMonitor {
