+++ src/modules/cbm/main.cpp
@@ -2,11 +2,11 @@
 //===----------------------------------------------------------------------===//
 ///
 /// \file
-/// \brief Main entry for ttlmonitor
+/// \brief Main entry for Common Beam Monitor (CBM) detector module.
 //===----------------------------------------------------------------------===//
 
 #include <efu/MainProg.h>
-#include <modules/ttlmonitor/TTLMonitorBase.h>
+#include <modules/cbm/TTLMonitorBase.h>
 
 int main(int argc, char *argv[]) {
   MainProg Main("ttlmonitor", argc, argv);
